Revert "Data constraints"
This reverts commit 326707267d1eed72b65db908902496a496debb88.

+++ assessment-backend/src/main/java/app/skillsoft/assessmentbackend/services/impl/CompetencyServiceImpl.java
@@ -94,7 +94,11 @@ public class CompetencyServiceImpl implements CompetencyService {
 
     @Override
     @Transactional
-    public void deleteCompetency(UUID id) {
-        competencyRepository.deleteById(id);
+    public boolean deleteCompetency(UUID id) {
+        if (competencyRepository.existsById(id)) {
+            competencyRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
